moved custom typings to src to allow for intellisense and added comments on go file functions
