+++ api/test.go
@@ -9,6 +9,7 @@ import (
 	"testing"
 )
 
+// PostFunc is a handler function which sends a POST request to the local API
 func PostFunc(t *testing.T, json string, url string, expectedCode int) {
 	os.Setenv("TEST", "1")
 	testRouter := SetupRouter(true, false)
@@ -26,6 +27,7 @@ func PostFunc(t *testing.T, json string, url string, expectedCode int) {
 	}
 }
 
+// GetFunc is a handler function which sends a GET request to the local API
 func GetFunc(t *testing.T, url string, expectedCode int) {
 	os.Setenv("TEST", "1")
 	testRouter := SetupRouter(true, false)
@@ -41,6 +43,7 @@ func GetFunc(t *testing.T, url string, expectedCode int) {
 	}
 }
 
+// DeleteFunc is a handler function which sends a DELETE request to the local API
 func DeleteFunc(t *testing.T, url string, expectedCode int) {
 	os.Setenv("TEST", "1")
 	testRouter := SetupRouter(true, false)
@@ -56,6 +59,7 @@ func DeleteFunc(t *testing.T, url string, expectedCode int) {
 	}
 }
 
+// SoftDeleteFunc is a handler function which sends a PATCH request to the local API
 func SoftDeleteFunc(t *testing.T, url string, expectedCode int) {
 	os.Setenv("TEST", "1")
 	testRouter := SetupRouter(true, false)
@@ -71,6 +75,7 @@ func SoftDeleteFunc(t *testing.T, url string, expectedCode int) {
 	}
 }
 
+// PutFunc is a handler function which sends a PUT request to the local API
 func PutFunc(t *testing.T, json string, url string, expectedCode int) {
 	os.Setenv("TEST", "1")
 	testRouter := SetupRouter(true, false)
